Deduplicate UI code
